fix bad syntax in vue files when style tags are used in SVG
prior to this change if there were style tags within an SVG, when we are building the vue file they would be copied over and embedded within the template. This is erroneous syntax in vue. This change will use regex to extract style tags and then put them at the end of the vue file after the template tags have already been inserted as they should be in a vue SFC.